KeyPresentWidget: Made it accept a key rather than a fingerprint
I'm torn on this one.  For one, we want thin interfaces, so a mere
string representing the key would be good. On the other hand, we want a
model, view, controller type of architecture where the view is a
relatively dumb entity. If the view needs to query the model (gpgmh)
then we're doing something wrong.